Add support for hwversion 3 (though default is still set to 2).  Many new features.
Added dynamic measurement of resistance across the velostat between channels, to compensate for error.
Added menu and various MIDI output settings, as well as a brightness control.
Added basic CAN bus support.  (I've tested messages can be sent and received, but we're not using it for anything yet.)
Added support for DAC5102a-based line out on hwversion 3 -- it outputs a sine wave on the tonic for now.
Implemented slow release of MPE pressure, so notes don't cut off instantly.